Application/players/requirements.txt
Flask==2.0.3
werkzeug==2.0.3
boto3==1.26.70
gunicorn==20.1.0
python-dotenv==1.0.0